Change Signature Pad
This commit changes the signature pad to a typed signature for ease of access when it comes to needing two signatures for two different owners. This also cleans up all the name input so that the form side looks cleaner. A disclaimer was added for the signatures, and hiding functionality was added to signature two, for when it is not needed. A p was added that reads the owners first name and last name and combines them. This is so that the user must type their name exactly correct for their signature to hold up legally.